requirements.txt
Flask
pypdf
google-generativeai